refactor(formApiClientFactory): rearrange types (#2061)
* refactor: shift EditFormFieldParams to types/api

* refactor: shift IPossiblyPrefilledField to types

* refactor: add DuplicateFormBody, FormUpdateParmas and PublicFormViewDto to types/api

* refactor: rreimport dependencies based on previous changes

* fix: remove CreateFormService left over from rebase
