+++ src/app/modules/form/admin-form/admin-form.types.ts
@@ -1,6 +1,5 @@
 import { Result } from 'neverthrow'
 
-import { EditFieldActions } from '../../../../shared/constants'
 import {
   IFieldSchema,
   IForm,
@@ -23,19 +22,6 @@ export type AssertFormFn = (
   form: IPopulatedForm,
 ) => Result<true, ForbiddenFormError>
 
-export type DuplicateFormBody = {
-  title: string
-} & (
-  | {
-      responseMode: ResponseMode.Email
-      emails: string | string[]
-    }
-  | {
-      responseMode: ResponseMode.Encrypt
-      publicKey: string
-    }
-)
-
 export type OverrideProps = {
   endPage?: IForm['endPage']
   startPage?: IForm['startPage']
@@ -44,36 +30,6 @@ export type OverrideProps = {
   responseMode: ResponseMode
   emails?: string | string[]
   publicKey?: string
-}
-
-export type EditFormFieldParams = {
-  field: IFieldSchema
-} & (
-  | {
-      action: {
-        name: Exclude<EditFieldActions, EditFieldActions.Reorder>
-      }
-    }
-  | {
-      action: {
-        name: EditFieldActions.Reorder
-        position: number
-      }
-    }
-)
-
-export type FormUpdateParams = {
-  editFormField?: EditFormFieldParams
-  authType?: IForm['authType']
-  emails?: IForm['emails']
-  esrvcId?: IForm['esrvcId']
-  form_logics?: IForm['form_logics']
-  hasCaptcha?: IForm['hasCaptcha']
-  inactiveMessage?: IForm['inactiveMessage']
-  permissionList?: IForm['permissionList']
-  status?: IForm['status']
-  title?: IForm['title']
-  webhook?: IForm['webhook']
 }
 
 export type EditFormFieldResult = Result<IFieldSchema[], EditFieldError>
