+++ src/app/modules/form/admin-form/admin-form.utils.ts
@@ -9,6 +9,7 @@ import {
   ResponseMode,
   Status,
 } from '../../../../types'
+import { DuplicateFormBody, EditFormFieldParams } from '../../../../types/api'
 import { createLoggerWithLabel } from '../../../config/logger'
 import { isPossibleEmailFieldSchema } from '../../../utils/field-validation/field-validation.guards'
 import {
@@ -38,8 +39,6 @@ import {
 } from './admin-form.errors'
 import {
   AssertFormFn,
-  DuplicateFormBody,
-  EditFormFieldParams,
   EditFormFieldResult,
   OverrideProps,
   PermissionLevel,
