+++ src/types/api/index.ts
@@ -1,5 +1,6 @@
 export * from './core'
 export * from './auth'
+export * from './field'
 export * from './form'
 export * from './billing'
 export * from './examples'
