+++ src/types/field/index.ts
@@ -30,6 +30,7 @@ export * from './homeNoField'
 export * from './imageField'
 export * from './longTextField'
 export * from './mobileField'
+export * from './myinfoField'
 export * from './nricField'
 export * from './numberField'
 export * from './radioField'
