+++ src/types/field/myinfoField.ts
@@ -1,0 +1,7 @@
+import { LeanDocument } from 'mongoose'
+
+import { IFieldSchema } from '../field'
+
+export interface IPossiblyPrefilledField extends LeanDocument<IFieldSchema> {
+  fieldValue?: string
+}
